+++ internal_response_writer.go
@@ -24,7 +24,6 @@ func (thd *thread) ReadFrom(r io.Reader) (int64, error) {
 
 type threadWithExtras interface {
 	Transaction
-	serverlessTransaction
 	internal.AddAgentAttributer
 	internal.AddAgentSpanAttributer
 }